Added the first half of a full user patch method to the server.
#story[899]
-Added a method to update a user's medications from the database to a
new list of medications. It does this in as few operations (database
calls) as possible, and edits existing entries where possible (instead
of removing and then adding an updated version)
-Added a similar method for procedures
-Added a similar method for diseases
-Added a method that updates attributes as well as uses these 3 new
update methods to update most user data with a single server api call.
Updates for waiting list items, organ donations, and history still need
to be implemented for this api call.