Reworked history to work with a client-server architecture. #story[873,899] #pair[ada130, nzo13]
-Modified the GET /users/:id endpoint so that it also fetches all user
history from the database and returns it in the JSON user object.
-Reworked history to be an attribute of users rather than being written
to a single file. It is stored as a list of history items.
-Reworked the user history display from parsing data from the history
file to working from the list of history items stored in a user instead.
It sorts the history items by day rather than by session like it was
done previously. The specific time and action/description for each
history item is shown in a tree view.
-Changed history calls to add history items to the current user rather
than writing them to file. More history calls need to be added.

-Fixed the Today and Organs Affected GUI items showing when viewing as a
user (they were meant to be hidden).
-Fixed the refresh button overlapping with the waiting list button.