Made user history work serverside, made history table update correctly. #story[873,899]
-Added history update functionality to the server API, so now patch
requests to the server will update the previous user history list to
a new version, in as few database operations as possible.

-Removed the old static History class and all usages, as it has now been
replaced.
-Made all history item updates also correctly update the user history
table on the GUI as soon as the updates happen.
-Added proper history logging for all disease related actions.
-Added a new isUniqueIdentifier for a user that already exists method
that uses the server. This allows for account settings changes.
-Fixed some setUser methods not automatically updating all gui elements
to show the user that was just set.
-Fixed text boxes deselecting if a mouse up event is detected outside
the text box.