Fixed empty preferred last name being null, worked on NHI tests. #story[1105]
-Fixed an issue where preferred last name was appearing as null for new
users, was an issue with how users were fetched from the database.
-Diagnosed issue where newly created users would not load, the issue is
that the server now requires a username to be supplied. Need to discuss
with team before making changes to this.
-Added tests for the new NHI method that does multiple validity checks
on NHI numbers.
-Added tests for some untested features related to user creation that
were encountered while working on NHI - user import (CSV and JSON), and
medications.
-Rearranged CSV/JSON import code to be neater and added some proper
test data for both.
-Removed ignored tests.